Starting work for transmittance on mesh

+++ src/core/scene/shape/cube.zig
@@ -66,7 +66,7 @@ pub const Cube = struct {
         return aabb.intersect(local_ray);
     }
 
-    pub fn visibility(ray: Ray, trafo: Trafo, entity: usize, filter: ?Filter, scene: *const Scene) ?Vec4f {
+    pub fn visibility(ray: Ray, trafo: Trafo, entity: u32, filter: ?Filter, scene: *const Scene) ?Vec4f {
         _ = ray;
         _ = trafo;
         _ = entity;
